+++ UnrealEngine/Engine/Plugins/Runtime/ChunkDownloader/Source/Private/PlatformStreamDownload.h
@@ -8,7 +8,7 @@ class FString;
 template <typename FuncType> class TFunction;
 
 typedef TFunction<void(int32 HttpStatus)> FDownloadComplete;
-typedef TFunction<void(int32 BytesReceived)> FDownloadProgress;
+typedef TFunction<void(uint64 BytesReceived)> FDownloadProgress;
 typedef TFunction<void(void)> FDownloadCancel;
 
 extern FDownloadCancel PlatformStreamDownload(const FString& Url, const FString& TargetFile, const FDownloadProgress& Progress, const FDownloadComplete& Callback);
